move gcc-specific signal handling out of main.c
