+++ src/CMakeLists.txt
@@ -167,6 +167,7 @@ set(SERVER_SRC
   bindings.c
   console.c
   helpers.c
+  signals.c
   main.c
   )
 
@@ -179,7 +180,9 @@ set (SERVER_SRC ${SERVER_SRC}
 endif(CURSES_FOUND)
 
 find_program(IWYU_PATH NAMES include-what-you-use iwyu)
-if(NOT IWYU_PATH)
+if(IWYU_PATH)
+  #  set(C_INCLUDE_WHAT_YOU_USE "${IWYU_PATH} -Xiwyu --no_fwd_decls")
+else(IWYU_PATH)
 	message(STATUS "Could not find the program include-what-you-use")
 endif()
