+++ src/signals.h
@@ -1,0 +1,4 @@
+#pragma once
+
+int setup_signal_handler(void);
+
